libnet
Sam's fork of libnet (bundled or not) is unfortunate useless to MinGW users. MSVC user are a bit better off, but it's still crappy. So on windows, user must link with my fork of libnet of build their own.
Having said that, my fork isn't really "ready", but so is this windows port. I'll provide a release soon and also add a notice in ettercap's README.PLATFORMS about this.