Even more desperate tweaks for missing setsid function
no setsid on windows. in addition to that the file hosting it (ec_daemon.c) seems abandoned. HAVE_DAEMON is not tested for and therefore NEVER true.

I feel this file should be entirely skipped on windows instead of adding yet another conditional around stuff. We don't "daemonize" things. We can't. Not like this at least

https://www.gnu.org/software/gnulib/manual/html_node/setsid.html