libettercap is trouble
Not sure if only on windows but libettercap needs some serious revision. Symbols seem to be exported at random and around 20 needed for linking with ettercap.exe and etterfilter.exe aren't exported at all.

That's actually the smaller problem. For now we simply tell libettercap to export all 557 symbols.

The bigger problem is that ettercap.exe segfaults at the very beginning of the main function. etterfilter.exe and etterlog.exe are perfectly fine. not sure who to blame libettercap or the daemon dilemma or something else entirely.

I will just get the rest of it to build (plugins, courses, gtk etc) and then consult the ettercap community, who sadly don't do windows :-(