HAVE_PLUGINS vs. ENABLE_PLUGINS
"ENABLE_PLUGINS" is a cached variable that is only used by cmake (cmake -DENABLE_PLUGINS .., if(ENABLE_PLUGINS) ect).
HAVE_PLUGINS is set as a preprocessor definition if the user choses to build the plugins alongside ettercap (the default).

This is my understanding of these two variable. However, they seem to be mixed up at times in CMakeLists.txt and EttercapLibCheck.cmake where HAVE_PLUGINS is treated as a cmake variable. This should be fixed. Strange that this works on linux but cmake on windows detects this and starts crying...