Improved GeoIP support
The cache variable HAVE_GEOIP is used wrongly. Users must set it to "on" to enable geoip support but it is also used by cmake to store the absolute or relative path to the geoip library.

The former happens on the command line (-DHAVE_GEOIP=on) and therefore overrides the latter. causing the build to fail because libettercap will be linked with the string "on" instead of libgeoip.a.

I think the check should be expanded in general. Camke should at least also look for GeoIP.h because if not, users may confuse it with geoip's successor geoip2. In fact, ettercap should be updated to support the newer geoip2 (library and database format).