Dropped regex in favor of pcreposix
The missing regex functions in PCRE's Windows port are provided by the static pcreposix library, which is build in addition to the pcre library when building on Windows. I didn't know that.

So on Windows, cmake will from now on also look for and link with pcreposix instead of GNU regex. This is much more convenient (PCRE's windows port is rock solid) and the right thing to do.