Don't build the curses interface by default on windows
It's a mess and although promising, fixing it is not the priority.

As stated in include\missing\ncurses_mingw.h:
"OS_WINDOWS doesn't really use the full ncurses yet, but a small layer on-top of PD-curses."

Right now, we have a mix of a old PDcurses library, a local ncurses_mingw.h (which is never included) some more local defines and functions that are now either part of the PDcurses library (vwprintw), have been declared deprecated by it (PDC_check_bios_key) and cmake eagerly looking for ncurses (and ncurses only).
Great.

The bad news is that PDcurses is much more limited then ncurses. The good news is that ncurses support on windows came a long way. On windows, standard ncurses may very well work. So there may not be any need for PDcurses and any of these "hacks".
But again, there's much more important things to worry about.

https://www.gnu.org/software/ncurses/#h4-port-mingw
http://invisible-island.net/ncurses/ncurses.html#download_mingw