Use "-Wl,--subsystem -Wl,windows" when linking executables (MInGW)
I was able to use cmake's cache variables ${CMAKE_CREATE_CONSOLE_EXE} and ${CMAKE_CREATE_WIN32_EXE} for this. Not anymore. No idea why. cmake, you're weird.
Personally, I don't see the the need for the this flag. Why is it used? 
The only purpose it seems to have is to bother and confuse me and cmake.